docs: Commented the functions and structures
