+++ discord/audit_log.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 )
 
+// AuditLog represents a list of administrative actions performed in a guild
+// https://discord.com/developers/docs/resources/audit-log#audit-log-object-audit-log-structure
 type AuditLog struct {
 	ApplicationCommands  []ApplicationCommand  `json:"application_commands"`
 	AuditLogEntries      []AuditLogEntry       `json:"audit_log_entries"`
@@ -16,6 +18,8 @@ type AuditLog struct {
 	Webhooks             []Webhook             `json:"webhooks"`
 }
 
+// AuditLogEntry represents an administrative action
+// https://discord.com/developers/docs/resources/audit-log#audit-log-entry-object-audit-log-entry-structure
 type AuditLogEntry struct {
 	TargetID   string                   `json:"target_id"`
 	Changes    []AuditLogChange         `json:"changes,omitempty"`
@@ -26,13 +30,16 @@ type AuditLogEntry struct {
 	Reason     string                   `json:"reason,omitempty"`
 }
 
+// AuditLogChange represents a potential change that has been performed during an administrative action (discord.AuditLogEntry)
+// https://discord.com/developers/docs/resources/audit-log#audit-log-change-object-audit-log-change-structure
 type AuditLogChange struct {
 	NewValue json.RawMessage `json:"new_value,omitempty"`
 	OldValue json.RawMessage `json:"old_value,omitempty"`
 	Key      string          `json:"key"`
 }
 
-func (c *AuditLogChange) ParseValuesTo(old, new interface{}) error {
+// ParseChangedValuesTo parses the different changed values in the audit log change (discord.AuditLogChange) to a custom structure
+func (c *AuditLogChange) ParseChangedValuesTo(old, new interface{}) error {
 	if err := json.Unmarshal(c.OldValue, &old); err != nil {
 		return fmt.Errorf("failed to unmarshal old value, %s", err)
 	}
@@ -42,6 +49,8 @@ func (c *AuditLogChange) ParseValuesTo(old, new interface{}) error {
 	return nil
 }
 
+// AuditLogActionType represents the type of administrative action (discord.AuditLogEntry) that has been performed
+// https://discord.com/developers/docs/resources/audit-log#audit-log-entry-object-audit-log-events
 type AuditLogActionType int
 
 const (
@@ -192,6 +201,8 @@ const (
 	AuditLogActionTypeAutoModerationUserCommunicationDisabled
 )
 
+// AuditLogEntryInformation represents additional information for some administrative actions (discord.AuditLogEntry)
+// https://discord.com/developers/docs/resources/audit-log#audit-log-entry-object-optional-audit-entry-info
 type AuditLogEntryInformation struct {
 	ApplicationID                 string `json:"application_id,omitempty"`
 	AutoModerationRuleName        string `json:"auto_moderation_rule_name,omitempty"`
