+++ discord/auto_moderation.go
@@ -1,5 +1,7 @@
 package discord
 
+// AutoModerationRule is a rule that trigger based on some criteria
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-rule-object-auto-moderation-rule-structure
 type AutoModerationRule struct {
 	ID              string                        `json:"id"`
 	GuildID         string                        `json:"guild_id"`
@@ -14,12 +16,16 @@ type AutoModerationRule struct {
 	ExemptChannels  []string                      `json:"exempt_channels"`
 }
 
+// AutoModerationEventType represents in what event context a rule (discord.AutoModerationRule) should be checked
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-rule-object-event-types
 type AutoModerationEventType int
 
 const (
-	AutoModerationEventTypeMessage AutoModerationEventType = 1 + iota
+	AutoModerationEventTypeMessageSend AutoModerationEventType = 1 + iota
 )
 
+// AutoModerationTriggerType represents the type of content which can trigger the rule (discord.AutoModerationRule)
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-rule-object-trigger-types
 type AutoModerationTriggerType int
 
 const (
@@ -30,6 +36,8 @@ const (
 	AutoModerationTriggerTypeMentionSpam
 )
 
+// AutoModerationTriggerMetadata represents additional data used to determine whether a rule (discord.AutoModerationRule) should be triggered
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-rule-object-trigger-metadata
 type AutoModerationTriggerMetadata struct {
 	KeywordFilter     []string                          `json:"keyword_filter"`
 	RegexPatterns     []string                          `json:"regex_patterns"`
@@ -38,6 +46,8 @@ type AutoModerationTriggerMetadata struct {
 	MentionTotalLimit int                               `json:"mention_total_limit"`
 }
 
+// AutoModerationKeywordPresetType represents the internally pre-defined word sets which will be searched for in content
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-rule-object-keyword-preset-types
 type AutoModerationKeywordPresetType int
 
 const (
@@ -46,11 +56,15 @@ const (
 	AutoModerationKeywordPresetTypeSlurs
 )
 
+// AutoModerationAction represents the action which will execute whenever a rule (discord.AutoModerationRule) is triggered
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-action-object-auto-moderation-action-structure
 type AutoModerationAction struct {
 	Type     AutoModerationActionType     `json:"type"`
 	Metadata AutoModerationActionMetadata `json:"metadata,omitempty"`
 }
 
+// AutoModerationActionType represents the type of action (discord.AutoModerationAction) that will be performed
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-action-object-action-types
 type AutoModerationActionType int
 
 const (
@@ -59,6 +73,8 @@ const (
 	AutoModerationActionTypeTimeout
 )
 
+// AutoModerationActionMetadata represents additional data used when an action (discord.AutoModerationAction) is executed. Different fields are relevant based on the value of action type (discord.AutoModerationActionType)
+// https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-action-object-action-metadata
 type AutoModerationActionMetadata struct {
 	ChannelID       string `json:"channel_id,omitempty"`
 	DurationSeconds int    `json:"duration_seconds,omitempty"`
