+++ discord/channel.go
@@ -1,11 +1,11 @@
 package discord
 
 import (
-	"encoding/json"
-	"fmt"
 	"time"
 )
 
+// Channel represents a guild or DM channel within Discord
+// https://discord.com/developers/docs/resources/channel#channel-object-channel-structure
 type Channel struct {
 	ID                            string                 `json:"id"`
 	Type                          ChannelType            `json:"type"`
@@ -44,6 +44,7 @@ type Channel struct {
 	DefaultForumLayout            DefaultForumLayoutView `json:"default_forum_layout,omitempty"`
 }
 
+// ChannelType represents the type of channel (discord.Channel)
 type ChannelType int
 
 const (
@@ -65,6 +66,8 @@ const (
 	ChannelTypeGuildForum
 )
 
+// Overwrite represents explicit permission overwrites for members (discord.Member) and roles (discord.Role)
+// https://discord.com/developers/docs/resources/channel#overwrite-object
 type Overwrite struct {
 	ID    string        `json:"id"`
 	Type  OverwriteType `json:"type"`
@@ -72,6 +75,7 @@ type Overwrite struct {
 	Deny  uint64        `json:"deny,string"`
 }
 
+// OverwriteType represents the type of overwrite (discord.Overwrite)
 type OverwriteType int
 
 const (
@@ -79,6 +83,8 @@ const (
 	OverwriteTypeMember
 )
 
+// VideoQualityMode represents the camera video quality mode of the voice channel (discord.Channel)
+// https://discord.com/developers/docs/resources/channel#channel-object-video-quality-modes
 type VideoQualityMode int
 
 const (
@@ -86,6 +92,8 @@ const (
 	VideoQualityModeFull
 )
 
+// ThreadMetadata represent additional information for thread channels (discord.Channel)
+// https://discord.com/developers/docs/resources/channel#thread-metadata-object-thread-metadata-structure
 type ThreadMetadata struct {
 	Archived            bool       `json:"archived"`
 	AutoArchiveDuration int        `json:"auto_archive_duration"`
@@ -95,6 +103,8 @@ type ThreadMetadata struct {
 	CreateTimestamp     time.Time  `json:"create_timestamp,omitempty"`
 }
 
+// ThreadMember represents a member that has joined a thread (discord.Channel & discord.ThreadMetadata)
+// https://discord.com/developers/docs/resources/channel#thread-member-object-thread-member-structure
 type ThreadMember struct {
 	ID            string     `json:"id,omitempty"`
 	UserID        string     `json:"user_id,omitempty"`
@@ -103,6 +113,8 @@ type ThreadMember struct {
 	Member        *Member    `json:"member,omitempty"`
 }
 
+// ChannelFlags represents the channel (discord.Channel) flags
+// https://discord.com/developers/docs/resources/channel#channel-object-channel-flags
 type ChannelFlags uint64
 
 const ChannelFlagsNone ChannelFlags = 0
@@ -114,6 +126,8 @@ const (
 	ChannelFlagRequireTag
 )
 
+// Tag represents a tag that is able to be applied to a thread in a forum channel (discord.ChannelTypeGuildForum)
+// https://discord.com/developers/docs/resources/channel#forum-tag-object-forum-tag-structure
 type Tag struct {
 	ID        string `json:"id"`
 	Name      string `json:"name"`
@@ -122,11 +136,15 @@ type Tag struct {
 	EmojiName string `json:"emoji_name,omitempty"`
 }
 
+// DefaultReaction represents the default reaction that is used when creating a new post in a forum channel (discord.ChannelTypeGuildForum)
+// https://discord.com/developers/docs/resources/channel#default-reaction-object-default-reaction-structure
 type DefaultReaction struct {
 	EmojiID   string `json:"emoji_id,omitempty"`
 	EmojiName string `json:"emoji_name,omitempty"`
 }
 
+// DefaultSortOrderType represents the default sorting method for forum channels (discord.ChannelTypeGuildForum)
+// https://discord.com/developers/docs/resources/channel#channel-object-sort-order-types
 type DefaultSortOrderType int
 
 const (
@@ -134,6 +152,8 @@ const (
 	DefaultSortOrderTypeCreationDate
 )
 
+// DefaultForumLayoutView represents the default layout view for forum channels (discord.ChannelTypeGuildForum)
+// https://discord.com/developers/docs/resources/channel#channel-object-forum-layout-types
 type DefaultForumLayoutView int
 
 const (
@@ -142,255 +162,16 @@ const (
 	DefaultForumLayoutViewGalleryView
 )
 
-type Message struct {
-	ID                   string                `json:"id"`
-	ChannelID            string                `json:"channel_id"`
-	Author               *User                 `json:"author"`
-	Content              string                `json:"content"`
-	Timestamp            time.Time             `json:"timestamp"`
-	EditedTimestamp      *time.Time            `json:"edited_timestamp,omitempty"`
-	TTS                  bool                  `json:"tts"`
-	MentionEveryone      bool                  `json:"mention_everyone"`
-	Mentions             []User                `json:"mentions"`
-	MentionRoles         []Role                `json:"mention_roles"`
-	MentionChannels      []Channel             `json:"mention_channels,omitempty"`
-	Attachments          []Attachment          `json:"attachments"`
-	Embeds               []Embed               `json:"embeds"`
-	Reactions            []Reaction            `json:"reactions,omitempty"`
-	Nonce                string                `json:"nonce,omitempty"`
-	Pinned               bool                  `json:"pinned"`
-	WebhookID            string                `json:"webhook_id,omitempty"`
-	Type                 MessageType           `json:"type"`
-	Activity             *MessageActivity      `json:"activity,omitempty"`
-	Application          *Application          `json:"application,omitempty"`
-	ApplicationID        string                `json:"application_id,omitempty"`
-	MessageReference     *MessageReference     `json:"message_reference,omitempty"`
-	Flags                MessageFlags          `json:"flags,omitempty"`
-	ReferencedMessage    *Message              `json:"referenced_message,omitempty"`
-	Interaction          *MessageInteraction   `json:"interaction,omitempty"`
-	Thread               *Channel              `json:"thread,omitempty"`
-	Components           []Component           `json:"components,omitempty"`
-	StickerItems         []StickerItem         `json:"sticker_items,omitempty"`
-	Stickers             []Sticker             `json:"stickers,omitempty"`
-	Position             int                   `json:"position"`
-	RoleSubscriptionData *RoleSubscriptionData `json:"role_subscription_data,omitempty"`
-}
-
-func (m *Message) UnmarshalJSON(b []byte) error {
-	type message Message
-	var source struct {
-		message
-		UnidentifiedComponents []UnidentifiedComponent `json:"components"`
-	}
-	if err := json.Unmarshal(b, &source); err != nil {
-		return fmt.Errorf("failed unmarshalling: %s", err)
-	}
-
-	*m = Message(source.message)
-
-	if len(source.UnidentifiedComponents) == 0 {
-		m.Components = nil
-		return nil
-	}
-
-	m.Components = make([]Component, len(source.UnidentifiedComponents))
-	for i, unidentifiedComponents := range source.UnidentifiedComponents {
-		m.Components[i] = unidentifiedComponents.data
-	}
-
-	return nil
-}
-
-type Attachment struct {
-	ID           string  `json:"id"`
-	Filename     string  `json:"filename"`
-	Description  string  `json:"description,omitempty"`
-	ContentType  string  `json:"content_type,omitempty"`
-	Size         int     `json:"size,omitempty"`
-	URL          string  `json:"url,omitempty"`
-	ProxyURL     string  `json:"proxy_url,omitempty"`
-	Height       int     `json:"height,omitempty"`
-	Width        int     `json:"width,omitempty"`
-	Ephemeral    bool    `json:"ephemeral,omitempty"`
-	DurationSecs float64 `json:"duration_secs,omitempty"`
-	Waveform     string  `json:"waveform,omitempty"`
-}
-
-type Embed struct {
-	Title       string         `json:"title,omitempty"`
-	Type        EmbedType      `json:"type,omitempty"`
-	Description string         `json:"description,omitempty"`
-	URL         string         `json:"url,omitempty"`
-	Timestamp   time.Time      `json:"timestamp,omitempty"`
-	Color       int            `json:"color,omitempty"`
-	Footer      *EmbedFooter   `json:"footer,omitempty"`
-	Image       *EmbedResource `json:"image,omitempty"`
-	Thumbnail   *EmbedResource `json:"thumbnail,omitempty"`
-	Video       *EmbedResource `json:"video,omitempty"`
-	Provider    *EmbedProvider `json:"provider,omitempty"`
-	Author      *EmbedAuthor   `json:"author,omitempty"`
-	Fields      []EmbedField   `json:"fields,omitempty"`
-}
-
-type EmbedType string
-
-const (
-	EmbedTypeRich    EmbedType = "rich"
-	EmbedTypeImage   EmbedType = "image"
-	EmbedTypVideo    EmbedType = "video"
-	EmbedTypeGIFV    EmbedType = "gifv"
-	EmbedTypeArticle EmbedType = "article"
-	EmbedTypeLink    EmbedType = "link"
-)
-
-type EmbedFooter struct {
-	Text         string `json:"text"`
-	IconURL      string `json:"icon_url,omitempty"`
-	ProxyIconURL string `json:"proxy_icon_url,omitempty"`
-}
-
-type EmbedResource struct {
-	Text         string `json:"text"`
-	ProxyIconURL string `json:"proxy_icon_url,omitempty"`
-	Height       int    `json:"height,omitempty"`
-	Width        int    `json:"width,omitempty"`
-}
-
-type EmbedProvider struct {
-	Name string `json:"name"`
-	URL  string `json:"url,omitempty"`
-}
-
-type EmbedAuthor struct {
-	Name         string `json:"name"`
-	URL          string `json:"url,omitempty"`
-	IconURL      string `json:"icon_url,omitempty"`
-	ProxyIconURL string `json:"proxy_icon_url,omitempty"`
-}
-
-type EmbedField struct {
-	Name   string `json:"name"`
-	Value  string `json:"value"`
-	Inline bool   `json:"inline,omitempty"`
-}
-
-type Reaction struct {
-	Count int   `json:"count"`
-	Me    bool  `json:"me"`
-	Emoji Emoji `json:"emoji"`
-}
-
-type Emoji struct {
-	ID            string   `json:"id"`
-	Name          string   `json:"name,omitempty"`
-	Roles         []string `json:"roles,omitempty"`
-	User          *User    `json:"user,omitempty"`
-	RequireColons bool     `json:"require_colons,omitempty"`
-	Managed       bool     `json:"managed,omitempty"`
-	Animated      bool     `json:"animated,omitempty"`
-	Available     bool     `json:"available,omitempty"`
-}
-
-type MessageType int
-
-const (
-	MessageTypeDefault MessageType = iota
-	MessageTypeRecipientAdd
-	MessageTypeRecipientRemove
-	MessageTypeCall
-	MessageTypeChannelNameChange
-	MessageTypeChannelIconChange
-	MessageTypeChannelPinnedMessage
-	MessageTypeUserJoin
-	MessageTypeGuildBoost
-	MessageTypeGuildBoostTier1
-	MessageTypeGuildBoostTier2
-	MessageTypeGuildBoostTier3
-	MessageTypeChannelFollowAdd
-	_
-	MessageTypeGuildDiscoveryDisqualified
-	MessageTypeGuildDiscoveryRequalified
-	MessageTypeGuildDiscoveryGracePeriodInitialWarning
-	MessageTypeGuildDiscoveryGracePeriodFinalWarning
-	MessageTypeThreadCreated
-	MessageTypeReply
-	MessageTypeChatInputCommand
-	MessageTypeThreadStarterMessage
-	MessageTypeGuildInviteReminder
-	MessageTypeContextMenuCommand
-	MessageTypeAutoModerationAction
-	MessageTypeRoleSubscriptionPurchase
-	MessageTypeInteractionPremiumUpsell
-	MessageTypeStageStart
-	MessageTypeStageEnd
-	MessageTypeStageSpeaker
-	_
-	MessageTypeStageTopic
-	MessageTypeGuildApplicationPremiumSubscription
-)
-
-type MessageActivity struct {
-	Type    MessageActivityType `json:"type"`
-	PartyID string              `json:"party_id,omitempty"`
-}
-
-type MessageActivityType int
-
-const (
-	MessageActivityTypeJoin MessageActivityType = 1 + iota
-	MessageActivityTypeSpectate
-	MessageActivityTypeListen
-	MessageActivityTypeJoinRequest
-)
-
-type MessageReference struct {
-	MessageID       string `json:"message_id"`
-	ChannelID       string `json:"channel_id,omitempty"`
-	GuildID         string `json:"guild_id,omitempty"`
-	FailIfNotExists bool   `json:"fail_if_not_exists,omitempty"`
-}
-
-type MessageFlags uint64
-
-const MessageFlagsNone MessageFlags = 0
-const (
-	MessageFlagCrossposted MessageFlags = 1 << iota
-	MessageFlagIsCrosspost
-	MessageFlagSuppressEmbeds
-	MessageFlagSourceMessageDeleted
-	MessageFlagUrgent
-	MessageFlagHasThread
-	MessageFlagEphemeral
-	MessageFlagLoading
-	MessageFlagFailedToMentionSomeRolesInThread
-	_
-	_
-	_
-	MessageFlagSuppressNotifications
-	MessageFlagIsVoiceMessage
-)
-
-type MessageInteraction struct {
-	ID     string          `json:"id"`
-	Type   InteractionType `json:"type"`
-	Name   string          `json:"name"`
-	User   *User           `json:"user"`
-	Member *Member         `json:"member,omitempty"`
-}
-
-type RoleSubscriptionData struct {
-	RoleSubscriptionListingID string `json:"role_subscription_listing_id"`
-	TierName                  string `json:"tier_name"`
-	TotalMonthsSubscribed     int    `json:"total_months_subscribed"`
-	IsRenewal                 bool   `json:"is_renewal"`
-}
-
+// ArchivedThreads represents the list of archived threads and members within those
+// https://discord.com/developers/docs/resources/channel#list-public-archived-threads-response-body
 type ArchivedThreads struct {
 	Threads []Channel      `json:"threads"`
 	Members []ThreadMember `json:"members"`
 	HasMore bool           `json:"has_more"`
 }
 
+// ActiveThreads represents the list of active threads and members within those
+// https://discord.com/developers/docs/resources/guild#list-active-guild-threads-response-body
 type ActiveThreads struct {
 	Threads []Channel      `json:"threads"`
 	Members []ThreadMember `json:"members"`
