+++ discord/guild_scheduled_event.go
@@ -2,6 +2,8 @@ package discord
 
 import "time"
 
+// GuildScheduledEvent represents a scheduled event in a guild (discord.Guild)
+// https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-structure
 type GuildScheduledEvent struct {
 	ID                 string                             `json:"id"`
 	GuildID            string                             `json:"guild_id"`
@@ -21,6 +23,8 @@ type GuildScheduledEvent struct {
 	Image              string                             `json:"image"`
 }
 
+// GuildScheduledEventPrivacyLevel represents the privacy level for the scheduled event (discord.GuildScheduledEvent)
+// https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-privacy-level
 type GuildScheduledEventPrivacyLevel int
 
 const (
@@ -28,6 +32,8 @@ const (
 	GuildScheduledEventPrivacyLevelGuildOnly
 )
 
+// GuildScheduledEventStatus represents the current status of the scheduled event (discord.GuildScheduledEvent)
+// https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-status
 type GuildScheduledEventStatus int
 
 const (
@@ -37,6 +43,8 @@ const (
 	GuildScheduledEventStatusCancelled
 )
 
+// GuildScheduledEventEntityType represents the type of the scheduled event (discord.GuildScheduledEvent)
+// https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-entity-types
 type GuildScheduledEventEntityType int
 
 const (
@@ -45,10 +53,14 @@ const (
 	GuildScheduledEventEntityTypeExternal
 )
 
+// GuildScheduledEventEntityMetadata represents additional metadata for the scheduled event (discord.GuildScheduledEvent) of type external (discord.GuildScheduledEventEntityTypeExternal)
+// https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-object-guild-scheduled-event-entity-metadata
 type GuildScheduledEventEntityMetadata struct {
 	Location string `json:"location,omitempty"`
 }
 
+// GuildScheduledEventUser represents a user (discord.User) and member (discord.Member), if any, that is subscribed to a scheduled event (discord.GuildScheduledEvent)
+// https://discord.com/developers/docs/resources/guild-scheduled-event#guild-scheduled-event-user-object-guild-scheduled-event-user-structure
 type GuildScheduledEventUser struct {
 	GuildScheduledEventID string  `json:"guild_scheduled_event_id"`
 	User                  *User   `json:"user"`
