+++ discord/interaction.go
@@ -1,5 +1,7 @@
 package discord
 
+// InteractionType represents the type of interaction that has been executed
+// https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-interaction-type
 type InteractionType int
 
 const (
