+++ discord/stage_instances.go
@@ -1,5 +1,7 @@
 package discord
 
+// StageInstance represents a live stage
+// https://discord.com/developers/docs/resources/stage-instance#stage-instance-object-stage-instance-structure
 type StageInstance struct {
 	ID                    string                    `json:"id"`
 	GuildID               string                    `json:"guild_id"`
@@ -10,6 +12,8 @@ type StageInstance struct {
 	GuildScheduledEventID string                    `json:"guild_scheduled_event_id"`
 }
 
+// StageInstancePrivacyLevel represents the privacy level of the stage instance (discord.StageInstance)
+// https://discord.com/developers/docs/resources/stage-instance#stage-instance-object-privacy-level
 type StageInstancePrivacyLevel int
 
 const (
