+++ discord/sticker.go
@@ -1,5 +1,7 @@
 package discord
 
+// Sticker represents a sticker that can be sent in messages
+// https://discord.com/developers/docs/resources/sticker#sticker-object-sticker-structure
 type Sticker struct {
 	ID          string            `json:"id"`
 	PackID      string            `json:"pack_id"`
@@ -14,6 +16,8 @@ type Sticker struct {
 	SortValue   int               `json:"sort_value"`
 }
 
+// StickerType represents the type of sticker (discord.Sticker)
+// https://discord.com/developers/docs/resources/sticker#sticker-object-sticker-types
 type StickerType int
 
 const (
@@ -21,6 +25,8 @@ const (
 	StickerTypeGuild
 )
 
+// StickerFormatType represents the type of format of the sticker (discord.Sticker)
+// https://discord.com/developers/docs/resources/sticker#sticker-object-sticker-format-types
 type StickerFormatType int
 
 const (
@@ -30,12 +36,16 @@ const (
 	StickerFormatTypeGIF
 )
 
+// StickerItem represents a partial sticker (discord.Sticker)
+// https://discord.com/developers/docs/resources/sticker#sticker-item-object-sticker-item-structure
 type StickerItem struct {
 	ID         string            `json:"id"`
 	Name       string            `json:"name"`
 	FormatType StickerFormatType `json:"format_type"`
 }
 
+// StickerPack represents a pack containing multiple standard stickers (discord.StickerTypeStandard)
+// https://discord.com/developers/docs/resources/sticker#sticker-pack-object-sticker-pack-structure
 type StickerPack struct {
 	ID             string    `json:"id"`
 	Stickers       []Sticker `json:"stickers"`
@@ -46,6 +56,8 @@ type StickerPack struct {
 	BannerAssetID  string    `json:"banner_asset_id"`
 }
 
+// NitroStickerPacks represents a list of default stickers (discord.Sticker) available for Nitro subscribers
+// https://discord.com/developers/docs/resources/sticker#list-nitro-sticker-packs-response-structure
 type NitroStickerPacks struct {
 	StickerPacks []StickerPack `json:"sticker_packs"`
 }
