+++ discord/voice.go
@@ -1,5 +1,7 @@
 package discord
 
+// VoiceRegion represents an available voice region
+// https://discord.com/developers/docs/resources/voice#voice-region-object-voice-region-structure
 type VoiceRegion struct {
 	ID         string `json:"id"`
 	Name       string `json:"name"`
