+++ oauth2/scopes.go
@@ -1,5 +1,7 @@
 package oauth2
 
+// Scope represents the various OAuth 2 scopes available
+// https://discord.com/developers/docs/topics/oauth2#shared-resources-oauth2-scopes
 type Scope string
 
 const (
