+++ rest/guild.go
@@ -10,7 +10,7 @@ const (
 	GuildsEndpoint = Endpoint + "guilds"
 )
 
-// GetGuild returns a discord.Guild based on its ID
+// GetGuild returns a guild structure (discord.Guild) for the given guild ID
 func (c *Client) GetGuild(guildID string, withCounts bool) (*discord.Guild, error) {
 	queryParams := make(QueryParameters)
 	if withCounts {
@@ -19,27 +19,27 @@ func (c *Client) GetGuild(guildID string, withCounts bool) (*discord.Guild, erro
 	return DoRequestAs[discord.Guild](c, "GET", GuildsEndpoint+"/"+guildID, queryParams, 1)
 }
 
-// GetGuildPreview returns a discord.GuildPreview based on its ID
+// GetGuildPreview returns a guild preview structure (discord.GuildPreview) for the given guild ID
 func (c *Client) GetGuildPreview(guildID string) (*discord.GuildPreview, error) {
 	return DoRequestAs[discord.GuildPreview](c, "GET", GuildsEndpoint+"/"+guildID+"/preview", nil, 1)
 }
 
-// GetGuildChannels returns a list of discord.Channel based on the guild ID
+// GetGuildChannels returns a list of channel structures (discord.Channel) for the given guild ID
 func (c *Client) GetGuildChannels(guildID string) ([]discord.Channel, error) {
 	return DoRequestAsList[discord.Channel](c, "GET", GuildsEndpoint+"/"+guildID+"/channels", nil, 1)
 }
 
-// ListActiveThreads returns a discord.ActiveThreads structure
+// ListActiveThreads returns an active threads structure (discord.ActiveThreads) for the given guild ID
 func (c *Client) ListActiveThreads(guildID string) (*discord.ActiveThreads, error) {
 	return DoRequestAs[discord.ActiveThreads](c, "GET", GuildsEndpoint+"/"+guildID+"/threads/active", nil, 1)
 }
 
-// GetGuildMember returns a discord.Member structure
+// GetGuildMember returns a member structure (discord.Member) for the given guild and user IDs
 func (c *Client) GetGuildMember(guildID, userID string) (*discord.Member, error) {
 	return DoRequestAs[discord.Member](c, "GET", GuildsEndpoint+"/"+guildID+"/members/"+userID, nil, 1)
 }
 
-// SearchGuildMember returns a list of discord.Member structure matching the given query
+// SearchGuildMember returns a list of member structures (discord.Member) matching the given query in the given guild ID
 func (c *Client) SearchGuildMember(guildID, query string, limit int) ([]discord.Member, error) {
 	queryParams := make(QueryParameters)
 	queryParams["query"] = query
@@ -49,7 +49,7 @@ func (c *Client) SearchGuildMember(guildID, query string, limit int) ([]discord.
 	return DoRequestAsList[discord.Member](c, "GET", GuildsEndpoint+"/"+guildID+"/members/search", queryParams, 1)
 }
 
-// GetGuildBans returns a list of discord.GuildBan structures
+// GetGuildBans returns a list of guild ban structures (discord.GuildBan) for the given guild ID
 func (c *Client) GetGuildBans(guildID, before, after string, limit int) ([]discord.GuildBan, error) {
 	queryParams := make(QueryParameters)
 	if before != "" {
@@ -64,17 +64,17 @@ func (c *Client) GetGuildBans(guildID, before, after string, limit int) ([]disco
 	return DoRequestAsList[discord.GuildBan](c, "GET", GuildsEndpoint+"/"+guildID+"/bans", queryParams, 1)
 }
 
-// GetGuildBan returns a discord.GuildBan structure matching the given ID
+// GetGuildBan returns a guild ban structure (discord.GuildBan) for the given guild and ban ID
 func (c *Client) GetGuildBan(guildID, banID string) (*discord.GuildBan, error) {
 	return DoRequestAs[discord.GuildBan](c, "GET", GuildsEndpoint+"/"+guildID+"/bans/"+banID, nil, 1)
 }
 
-// GetGuildRoles returns a list of discord.Role structures
+// GetGuildRoles returns a list role structures (discord.Role)
 func (c *Client) GetGuildRoles(guildID string) ([]discord.Role, error) {
 	return DoRequestAsList[discord.Role](c, "GET", GuildsEndpoint+"/"+guildID+"/roles", nil, 1)
 }
 
-// GetGuildPruneCount returns a discord.GuildPrune structure
+// GetGuildPruneCount returns a guild prune structure (discord.GuildPrune) for the given guild IDs
 func (c *Client) GetGuildPruneCount(guildID string, days int, includeRoles []string) (*discord.GuildPrune, error) {
 	queryParams := make(QueryParameters)
 	if days >= 1 && days <= 30 {
@@ -86,42 +86,42 @@ func (c *Client) GetGuildPruneCount(guildID string, days int, includeRoles []str
 	return DoRequestAs[discord.GuildPrune](c, "GET", GuildsEndpoint+"/"+guildID+"/prune", queryParams, 1)
 }
 
-// GetGuildVoiceRegions returns a list of discord.VoiceRegion structures
+// GetGuildVoiceRegions returns a list of voice region structures (discord.VoiceRegion) for the given guild ID
 func (c *Client) GetGuildVoiceRegions(guildID string) ([]discord.VoiceRegion, error) {
 	return DoRequestAsList[discord.VoiceRegion](c, "GET", GuildsEndpoint+"/"+guildID+"/regions", nil, 1)
 }
 
-// GetGuildInvites returns a list of discord.InviteWithMetadata structures
+// GetGuildInvites returns a list of invite with metadata structures (discord.InviteWithMetadata) for the given guild ID
 func (c *Client) GetGuildInvites(guildID string) ([]discord.InviteWithMetadata, error) {
 	return DoRequestAsList[discord.InviteWithMetadata](c, "GET", GuildsEndpoint+"/"+guildID+"/invites", nil, 1)
 }
 
-// GetGuildIntegrations returns a list of discord.Integration structures
+// GetGuildIntegrations returns a list of integrations structures (discord.Integration) for the given guild ID
 func (c *Client) GetGuildIntegrations(guildID string) ([]discord.Integration, error) {
 	return DoRequestAsList[discord.Integration](c, "GET", GuildsEndpoint+"/"+guildID+"/integrations", nil, 1)
 }
 
-// GetGuildWidgetSetting returns a discord.WidgetSetting structure
+// GetGuildWidgetSetting returns a widget setting structure (discord.WidgetSetting) for the given guild ID
 func (c *Client) GetGuildWidgetSetting(guildID string) (*discord.WidgetSetting, error) {
 	return DoRequestAs[discord.WidgetSetting](c, "GET", GuildsEndpoint+"/"+guildID+"/widget", nil, 1)
 }
 
-// GetGuildWidget returns a discord.Widget structure
+// GetGuildWidget returns a widget structure (discord.Widget) for the given guild ID
 func (c *Client) GetGuildWidget(guildID string) (*discord.Widget, error) {
 	return DoRequestAs[discord.Widget](c, "GET", GuildsEndpoint+"/"+guildID+"/widget.json", nil, 1)
 }
 
-// GetGuildVanityURL returns a discord.PartialInvite structure
+// GetGuildVanityURL returns the vanity invite, which is a partial invite structure (discord.PartialInvite), for the given guild ID
 func (c *Client) GetGuildVanityURL(guildID string) (*discord.PartialInvite, error) {
 	return DoRequestAs[discord.PartialInvite](c, "GET", GuildsEndpoint+"/"+guildID+"/vanity-url", nil, 1)
 }
 
-// GetGuildWelcomeScreen returns a discord.WelcomeScreen structure
+// GetGuildWelcomeScreen returns a welcome screen structure (discord.WelcomeScreen) for the given guild ID
 func (c *Client) GetGuildWelcomeScreen(guildID string) (*discord.WelcomeScreen, error) {
 	return DoRequestAs[discord.WelcomeScreen](c, "GET", GuildsEndpoint+"/"+guildID+"/welcome-screen", nil, 1)
 }
 
-// GetGuildOnboarding returns a discord.Onboarding structure
+// GetGuildOnboarding returns a onboarding structure (discord.Onboarding) for the given guild ID
 func (c *Client) GetGuildOnboarding(guildID string) (*discord.Onboarding, error) {
 	return DoRequestAs[discord.Onboarding](c, "GET", GuildsEndpoint+"/"+guildID+"/onboarding", nil, 1)
 }
